auth restructure
