+++ src/auth/auth_manager.rs
@@ -1,74 +1,9 @@
 use crate::session::Session;
-use argon2::{
-    password_hash::{Error, PasswordHash, PasswordHasher, PasswordVerifier, SaltString},
-    Argon2,
-};
+
 use regex::Regex;
 use std::collections::HashMap;
 
-// Permissions bit mask:
-// 0b00000001 - SET
-// 0b00000010 - GET
-// 0b00000100 - DEL
-// 0b00001000 - CREATE_USER & DELETE_USER
-// To GRANT permission user needs 0b00001000 & appropriate permission:
-// 0b00001000 | 0b00000001 = 0b00001001
-// 0b00001000 | 0b00000010 = 0b00001010
-// 0b00001000 | 0b00000100 = 0b00001100
-
-#[allow(non_camel_case_types)]
-#[derive(Debug, PartialEq)]
-pub enum Permissions {
-    NONE = 0,
-    SET = 1 << 0,
-    GET = 1 << 1,
-    DEL = 1 << 2,
-    USER_ADMIN = 1 << 3,
-}
-
-impl Permissions {
-    pub fn from_u8(value: u8) -> Vec<Permissions> {
-        let mut permissions = Vec::new();
-
-        if value & (Permissions::SET as u8) != 0 {
-            permissions.push(Permissions::SET);
-        }
-
-        if value & (Permissions::GET as u8) != 0 {
-            permissions.push(Permissions::GET);
-        }
-
-        if value & (Permissions::DEL as u8) != 0 {
-            permissions.push(Permissions::DEL);
-        }
-
-        if value & (Permissions::USER_ADMIN as u8) != 0 {
-            permissions.push(Permissions::USER_ADMIN);
-        }
-
-        permissions
-    }
-}
-
-#[derive(Debug)]
-pub struct User {
-    username: String,
-    password: String,
-    permissions: u8,
-}
-
-impl User {
-    pub fn new(username: String, password: String, permissions: u8) -> User {
-        User {
-            username,
-            password,
-            permissions,
-        }
-    }
-    pub fn to_string(&self) -> String {
-        format!("User: {} Permissions: {}", self.username, self.permissions)
-    }
-}
+use crate::auth::{Permissions, User};
 
 #[derive(Debug)]
 pub struct AuthManager {
@@ -116,26 +51,6 @@ impl AuthManager {
         Ok(())
     }
 
-    fn hash_password(password: &str) -> Result<String, Error> {
-        let argon2 = Argon2::default();
-        let salt = SaltString::generate(&mut rand::thread_rng());
-        let hash = argon2
-            .hash_password(password.as_bytes(), &salt)?
-            .to_string();
-
-        Ok(hash)
-    }
-
-    fn verify_password(hash: &str, password: &str) -> Result<bool, Error> {
-        let argon2 = Argon2::default();
-        let parsed_hash = PasswordHash::new(hash)?;
-
-        match argon2.verify_password(password.as_bytes(), &parsed_hash) {
-            Ok(_) => Ok(true),
-            Err(_) => Ok(false),
-        }
-    }
-
     pub fn create_user(
         &mut self,
         username: String,
@@ -144,12 +59,16 @@ impl AuthManager {
     ) -> Result<String, String> {
         Self::validate_password(&password)?;
 
-        let hash = Self::hash_password(&password).map_err(|e| e.to_string())?;
+        if self.users.contains_key(&username) {
+            return Err("User already exists".to_string());
+        }
 
-        self.users.insert(
-            username.clone(),
-            User::new(username.clone(), hash, permission),
-        );
+        let new_user = match User::new(username.clone(), password.clone(), permission) {
+            Ok(user) => user,
+            Err(e) => return Err(e.to_string()),
+        };
+
+        self.users.insert(username.clone(), new_user);
 
         Ok("OK".to_string())
     }
@@ -175,7 +94,7 @@ impl AuthManager {
         match self.users.get(&username) {
             None => return Err("Username or password is incorrect".to_string()),
             Some(user) => {
-                if Self::verify_password(&user.password, &password).map_err(|e| e.to_string())? {
+                if user.verify_password(&password).map_err(|e| e.to_string())? {
                     Ok(session.set_authenticated(&username))
                 } else {
                     Err("Username or password is incorrect".to_string())
@@ -208,10 +127,8 @@ impl AuthManager {
             None => return Err("User does not exist".to_string()),
             Some(user) => {
                 let new_permissions = user.permissions | permission;
-                self.users.insert(
-                    username.clone(),
-                    User::new(username.clone(), user.password.clone(), new_permissions),
-                );
+                self.users
+                    .insert(username.clone(), user.update_permissions(new_permissions));
                 Ok("OK".to_string())
             }
         }
@@ -226,10 +143,8 @@ impl AuthManager {
             None => return Err("User does not exist".to_string()),
             Some(user) => {
                 let new_permissions = user.permissions & !permission;
-                self.users.insert(
-                    username.clone(),
-                    User::new(username, user.password.clone(), new_permissions),
-                );
+                self.users
+                    .insert(username.clone(), user.update_permissions(new_permissions));
                 Ok("OK".to_string())
             }
         }
