+++ src/auth/mod.rs
@@ -1,0 +1,10 @@
+mod auth_manager;
+mod permission;
+mod user;
+
+pub use auth_manager::*;
+pub use permission::*;
+pub use user::*;
+
+#[cfg(test)]
+mod test;
