+++ src/auth/test/mod.rs
@@ -1,0 +1,1 @@
+mod auth_manager_tests;
