+++ src/data/mod.rs
@@ -1,10 +1,8 @@
-mod auth_manager;
 mod data_manager;
 mod data_type;
 mod data_value;
 mod key;
 mod store;
-pub use auth_manager::*;
 pub use data_manager::*;
 pub use data_type::*;
 pub use store::Store;
