+++ src/data/test/mod.rs
@@ -1,4 +1,3 @@
-mod auth_manager_tests;
 mod command_tests;
 mod data_tests_utils;
 mod key_tests;
