+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod auth;
 pub mod commands;
 pub mod data;
 pub mod handler;
